test(raft): clean up tests
