+++ protocols/raft/src/test/java/io/atomix/protocols/raft/primitive/TestPrimitiveClient.java
@@ -1,0 +1,17 @@
+package io.atomix.protocols.raft.primitive;
+
+import io.atomix.primitive.event.Event;
+
+/**
+ * Test primitive client interface.
+ */
+public interface TestPrimitiveClient {
+  @Event("event")
+  void event(long index);
+
+  @Event("expire")
+  void expire(String value);
+
+  @Event("close")
+  void close(String value);
+}
