+++ protocols/raft/src/test/java/io/atomix/protocols/raft/session/impl/RaftSessionStateTest.java
@@ -18,7 +18,7 @@ package io.atomix.protocols.raft.session.impl;
 import java.util.UUID;
 
 import io.atomix.primitive.session.SessionId;
-import io.atomix.protocols.raft.TestPrimitiveType;
+import io.atomix.protocols.raft.session.TestPrimitiveType;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
